+++ bindings/nostr-sdk-ffi/src/pool/result.rs
@@ -9,17 +9,19 @@ use nostr_ffi::EventId;
 use nostr_sdk::pool;
 use uniffi::Record;
 
-/// Send output
+/// Output
+///
+/// Send or negentropy reconciliation output
 #[derive(Record)]
-pub struct SendOutput {
-    /// Set of relay urls to which the message/s was successfully sent
+pub struct Output {
+    /// Set of relays that success
     pub success: Vec<String>,
-    /// Map of relay urls with related errors where the message/s wasn't sent
+    /// Map of relays that failed, with related errors.
     pub failed: HashMap<String, Option<String>>,
 }
 
-impl From<pool::SendOutput> for SendOutput {
-    fn from(value: pool::SendOutput) -> Self {
+impl From<pool::Output> for Output {
+    fn from(value: pool::Output) -> Self {
         Self {
             success: value.success.into_iter().map(|u| u.to_string()).collect(),
             failed: value
@@ -36,44 +38,15 @@ impl From<pool::SendOutput> for SendOutput {
 pub struct SendEventOutput {
     /// Event ID
     pub id: Arc<EventId>,
-    /// Set of relay urls to which the message/s was successfully sent
-    pub success: Vec<String>,
-    /// Map of relay urls with related errors where the message/s wasn't sent
-    pub failed: HashMap<String, Option<String>>,
+    /// Output
+    pub output: Output,
 }
 
 impl From<pool::SendEventOutput> for SendEventOutput {
     fn from(value: pool::SendEventOutput) -> Self {
         Self {
             id: Arc::new(value.id.into()),
-            success: value.success.into_iter().map(|u| u.to_string()).collect(),
-            failed: value
-                .failed
-                .into_iter()
-                .map(|(u, e)| (u.to_string(), e))
-                .collect(),
-        }
-    }
-}
-
-/// Negentropy reconciliation output
-#[derive(Record)]
-pub struct ReconciliationOutput {
-    /// Set of relay urls to which the negentropy reconciliation success
-    pub success: Vec<String>,
-    /// Map of relay urls with related errors where the negentropy reconciliation failed
-    pub failed: HashMap<String, Option<String>>,
-}
-
-impl From<pool::ReconciliationOutput> for ReconciliationOutput {
-    fn from(value: pool::ReconciliationOutput) -> Self {
-        Self {
-            success: value.success.into_iter().map(|u| u.to_string()).collect(),
-            failed: value
-                .failed
-                .into_iter()
-                .map(|(u, e)| (u.to_string(), e))
-                .collect(),
+            output: value.output.into(),
         }
     }
 }