+++ crates/nostr-sdk/examples/client.rs
@@ -21,10 +21,10 @@ async fn main() -> Result<()> {
     client.connect().await;
 
     // Publish a text note
-    let output = client.publish_text_note("Hello world", []).await?;
-    println!("Event ID: {}", output.id.to_bech32()?);
+    let SendEventOutput { id, output } = client.publish_text_note("Hello world", []).await?;
+    println!("Event ID: {}", id.to_bech32()?);
     println!("Sent to: {:?}", output.success);
-    // println!("Not set to: {:?}", output.failed);
+    println!("Not sent to: {:?}", output.failed);
 
     // Create a text note POW event and broadcast to all connected relays
     let event: Event =