+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * pool: use per-purpose dedicated relay channels ([Yuki Kishimoto])
 * pool: return relay urls to which `messages`/`events` have or not been sent for `send_*` and `batch_*` methods ([Yuki Kishimoto])
 * pool: rename `Relay::terminate` to `Relay::disconnect` ([Yuki Kishimoto])
+* pool: merge `SendOutput` and `ReconciliationOutput` in a single `Output` struct ([Yuki Kishimoto])
 * ffi(sdk): convert `RelayPool::handle_notifications` method to async/future ([Yuki Kishimoto])
 * js: increase max stack size to `0x1E84800` bytes (32 MiB) ([Yuki Kishimoto])